Enable bugprone-implicit-widening-of-multiplication-result
224 sites over all 840 translation units, cleared with no suppressions and no
check options: the tree measures 0 with the check at its defaults. Every one is
an int-by-int product that was widened afterwards to a 64-bit size, pointer
offset or timer value; the fix is always to perform the multiplication in the
destination's width instead.

Four shapes, and the leverage was in the types rather than the expressions.

Framebuffer strides. Six local declarations in src/sdllib/drawbuff.cc changed
from int to base::ssize -- the stride and area variables -- took that file from
23 sites to 8 on their own, because the row arithmetic then happens in 64-bit
at the source. Same in iconcach.cc, gbuffer.h and tech/2keyfbuf.cc.

Timer durations, which were most of the tree. TICKS_PER_MINUTE and TIMER_MINUTE
hide their multiplication inside the macro, so those sites cannot be fixed where
they appear -- the line just reads TICKS_PER_MINUTE. Widening the three derived
macros cleared 43 sites at once. The base rates TICKS_PER_SECOND and
TIMER_SECOND deliberately stay int: they are also divided into int quantities
(FADE_PALETTE_*, framedelay, fixed()), and widening them would push implicit
narrowing into a dozen contexts where clang-diagnostic-shorten-64-to-32 is still
disabled and would not catch it. The 34 remaining use sites got
int64_t{TICKS_PER_SECOND} * N instead.

Constant sizes. 64L * 1024 * 1024, 16UL * 1024 * sizeof(void*), 256UL * 3 --
the literal carries the width, which is the honest fix for a compile-time
constant rather than wrapping the product in a cast.

Everything else -- palette strides, icon strides, buffer allocations -- got an
explicit cast on the left operand, in the destination's type: base::ssize for
pointer offsets and sizes, std::size_t or long where the destination is
unsigned or a C library parameter.

Do not trust clang-tidy's fix-it here. Both of the notes it offers, including
the one labelled "perform multiplication in a wider type", carry replacements
spanning the whole product: applying them yields
static_cast<base::ssize>(y * dst_area), which leaves the multiplication in
32 bits and only makes the widening explicit. That is the silencing fix wearing
the correct fix's label. The usable signal is the note's Ranges entry, which
covers just the left operand; wrapping that is the real fix, and the 180
insertions applied that way were derived from it. Two further traps: clang
names some destination types with internal aliases (__size_t, __ptrdiff_t) that
cannot be written in source, and exporting fixes keyed by basename silently
loses files, since ra/house.cc and td/house.cc collide.

One retype bit back and is left in place deliberately. MAX_MESSAGE_LENGTH
became std::size_t -- it is a buffer length feeding port::SafeAppend -- which
broke std::min(maxchars, MAX_MESSAGE_LENGTH) in src/td/msglist.cc. The
comparison is now explicit rather than the macro reverted.

Both games build, the 84 tests pass, rasdl runs the intro against the Steam
data, and a full-config sweep over all 840 units reports 0 findings and no hard
error: lines.

+++ src/ra/cell.cc
@@ -2123,7 +2123,7 @@ long CellClass::Tiberium_Adjust(bool pregame) {
       } else {
         OverlayData = _adj[count];
       }
-      return (OverlayData + 1) * value;
+      return (static_cast<long>(OverlayData + 1)) * value;
     }
   }
   return 0;
